requirements.txt
# Job Automation System - Dependencies

# Email
imaplib2>=3.6

# AI / LLM
openai>=1.0
requests>=2.31
google-generativeai>=0.8.0
anthropic>=0.80.0
groq>=1.0.0

# Document parsing & generation
python-docx>=1.1
PyPDF2>=3.0
fpdf2>=2.7
PyYAML>=6.0

# HTML parsing
beautifulsoup4>=4.12
lxml>=5.0

# Browser automation
selenium==4.22.0
undetected-chromedriver>=3.5
selenium-stealth>=1.0.6
webdriver-manager>=4.0

# CLI & display
rich>=13.0

# Configuration
python-dotenv>=1.0

# Scheduling
schedule>=1.2

# Fuzzy matching (for job role filtering)
thefuzz>=0.22
python-Levenshtein>=0.25

# GUI
customtkinter==5.2.2
pillow==11.1.0
pyinstaller>=6.0.0